test(mod_arith): widen interpreter test storage to i16
fixup for the verifier commit: !mod_arith.int<251 : i8> violates the new
storage-one-bit-wider-than-modulus invariant (251 needs 8 bits, so i8
storage is rejected). Use i16 storage instead.

+++ Test/Interpreter/ModArith/mul.mlir
@@ -1,12 +1,12 @@
 // RUN: veir-interpret %s | filecheck %s
 
 "builtin.module"() ({
-  "func.func"() <{sym_name = "main", function_type = () -> !mod_arith.int<251 : i8>}> ({
-    %lhs = "mod_arith.constant"() <{ "value" = 250 : i8 }> : () -> !mod_arith.int<251 : i8>
-    %rhs = "mod_arith.constant"() <{ "value" = 250 : i8 }> : () -> !mod_arith.int<251 : i8>
-    %product = "mod_arith.mul"(%lhs, %rhs) : (!mod_arith.int<251 : i8>, !mod_arith.int<251 : i8>) -> !mod_arith.int<251 : i8>
-    "func.return"(%product) : (!mod_arith.int<251 : i8>) -> ()
+  "func.func"() <{sym_name = "main", function_type = () -> !mod_arith.int<251 : i16>}> ({
+    %lhs = "mod_arith.constant"() <{ "value" = 250 : i16 }> : () -> !mod_arith.int<251 : i16>
+    %rhs = "mod_arith.constant"() <{ "value" = 250 : i16 }> : () -> !mod_arith.int<251 : i16>
+    %product = "mod_arith.mul"(%lhs, %rhs) : (!mod_arith.int<251 : i16>, !mod_arith.int<251 : i16>) -> !mod_arith.int<251 : i16>
+    "func.return"(%product) : (!mod_arith.int<251 : i16>) -> ()
   }) : () -> ()
 }) : () -> ()
 
-// CHECK: Program output: #[0x01#8]
+// CHECK: Program output: #[0x0001#16]
